Fixed one nasty bug and many other syscall improvements
First the nasty bug: since I am creating a single Net::Graphite object
and reusing it for the lifetime of the application the data in the
object does not get cleared "implicitly", instead I expected the object
to do the right thing.
However, every time I called ->send with complex data structures it
would append the metrics to the _flush_buffer array and ->flush would
never clear that buffer. This meant that every time I called ->send it
would not only send this data, but also all the previous data points I
generated. After a short time the application would use huge amounts of
memory and cpu basically resending the same data.

Improvements:
I have reduced the number of syscalls that the module is making. Before
the change it would call getpeername() and sendto() syscall for every
single metric that it was sending to graphite. I have improved it in two
ways: 1. build a buffer with multiple metrics to send at once, rather
than one-at-a-time; 2. I have replaced IO::Socket->send method with a
direct call to CORE::send() perl function. The IO::Socket implementation
would call getpeername() every time for no good reason (it should probably
cache the result after the first call). Calling send() directly further
halved the number of syscalls.

I have also removed the ->connected check. This one again would call
getpeername() to check whether the connection is still alive. However, a
successfull check does not guarantee that we will be able to ->send
anything, so we have to check for the errors there anyway. So instead of
checking for connection we just try to send the data directly and if that
fails we will reconnect.